requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
billiard==3.6.4.0
bitarray==1.2.2
boto3==1.17.70
botocore==1.20.70
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
cytoolz==0.11.0
defusedxml==0.7.1
Django==3.2.2
django-admin-logs==1.0.2
django-allauth==0.44.0
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-magiclink==1.0.4
django-redis==5.0.0
django-reversion==3.0.9
django-storages==1.11.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
filetype==1.0.7
gunicorn==19.9.0
hexbytes==0.2.1
idna==2.10
iniconfig==1.1.1
ipfs-api==0.2.3
ipfshttpclient==0.7.0
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.1.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
oauthlib==3.1.1
packaging==20.9
parsimonious==0.8.1
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-django==4.4.0
pytest-mock==3.6.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rlp==2.0.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
toolz==0.11.1
typing-extensions==3.10.0.0
urllib3==1.26.4
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.20.0
websockets==8.1
yarl==1.6.3
psycopg2-binary>=2.8.6
django-health-check>=3.16.4
